fix: expose drawing options as function params rather than env config
